+++ NEWS.md
@@ -6,3 +6,9 @@
 - Provides model selection based on log-likelihood criteria.
 
 * Initial CRAN submission.
+
+# splineCox 1.0.2
+
+- Fixed bugs in the `splineCox.reg2` function when handling custom numeric vectors.
+- Added examples for custom vector normalization in the vignette.
+- Improved documentation and error messages.